% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FizzBuzz.R
\name{FizzBuzz}
\alias{FizzBuzz}
\title{FizzBuzz}
\usage{
FizzBuzz(n)
}
\arguments{
\item{n}{A vector of natural numbers.}
}
\value{
A character vector of numbers and words.
}
\description{
Takes a vector and returns "Fizz Buzz" if the element
is divisible by 3 and 5, "Fizz" if the element is divisible
by 3 and "Buzz" if the element is divisible by 5.
}
